Moving rabbitmq module into queue package
