+++ src/app/beer_garden/instances.py
@@ -10,7 +10,7 @@ from brewtils.models import Events, Instance, PatchOperation, System
 import beer_garden
 import beer_garden.db.api as db
 from beer_garden.events import publish_event
-from beer_garden.rabbitmq import get_routing_key, get_routing_keys
+from beer_garden.queue.rabbitmq import get_routing_key, get_routing_keys
 
 logger = logging.getLogger(__name__)
