+++ src/app/beer_garden/local_plugins/manager.py
@@ -7,7 +7,7 @@ import beer_garden
 import beer_garden.db.api as db
 from beer_garden.errors import PluginStartupError
 from beer_garden.local_plugins.plugin_runner import LocalPluginRunner
-from beer_garden.rabbitmq import get_routing_key
+from beer_garden.queue.rabbitmq import get_routing_key
 
 
 class LocalPluginsManager(object):
