+++ src/app/beer_garden/systems.py
@@ -11,7 +11,7 @@ from time import sleep
 import beer_garden
 import beer_garden.db.api as db
 from beer_garden.events import publish_event
-from beer_garden.rabbitmq import get_routing_key
+from beer_garden.queue.rabbitmq import get_routing_key
 
 REQUEST_FIELDS = set(SystemSchema.get_attribute_names())
